python_ocr_service/requirements.txt
Flask==2.3.3
Pillow==10.0.1
easyocr==1.7.0
opencv-python==4.8.1.78
numpy==1.24.3
google-generativeai==0.3.2
requests==2.31.0